+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+ - Fixed the `Mapping` objects not implementing `Sync`.
+
 ## Version 0.3.1 (2015-04-18)
 
  - Fixed lifetime-related issues when using the `uniform!` macro.