+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
  - Added `vertex::EmptyVertexAttributes` and `vertex::EmptyInstanceAttributes` markers in order to draw without a vertex source.
+ - Added more formats in `CompressedFormats` and `CompressedSrgbFormats`.
  - Fixed the `Mapping` objects not implementing `Sync`.
  - Fixed rendering to an sRGB texture not possible.
  - `GliumCreationError` now has a template parameter indicating the backend creation error type.